+++ src/sdk/PnP.Core.Admin/Model/Microsoft365/Public/Enums/GeoLocation.cs
@@ -83,6 +83,16 @@
         /// <summary>
         /// United Kingdom
         /// </summary>
-        GBR
+        GBR,
+
+        /// <summary>
+        /// Qatar
+        /// </summary>
+        QAT,
+
+        /// <summary>
+        /// Sweden
+        /// </summary>
+        SWE,
     }
 }