+++ src/sdk/PnP.Core.Admin/Model/Microsoft365/Public/Options/GraphGroupOptions.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Text.Json.Serialization;
 
 namespace PnP.Core.Admin.Model.Microsoft365
@@ -65,5 +64,11 @@ namespace PnP.Core.Admin.Model.Microsoft365
         /// If it is a security enabled group
         /// </summary>
         public bool SecurityEnabled { get; } = false;
+
+        /// <summary>
+        /// Allows defining the resource behavior options for the group
+        /// See https://learn.microsoft.com/en-us/graph/group-set-options#configure-groups
+        /// </summary>
+        public List<string> ResourceBehaviorOptions { get; set; }
     }
 }