+++ src/sdk/CHANGELOG.md
@@ -17,6 +17,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - `IPnPContext` interface on `PnPContext` to support mocking for use in testing #948 [jansenbe - Bert Jansen]
 - Support for getting, setting and clearing of (folder based) column defaults #949 [jansenbe - Bert Jansen]
 - Requesting re-indexing of a web or list #950 [jansenbe - Bert Jansen]
+- RateLimit header processing capability, turned off by default [jansenbe - Bert Jansen]
+- Added `DeleteRecycledSiteCollection` methods to delete site collections from the recycle bin [jansenbe - Bert Jansen]
+- Admin library: ability to specify ResourceBehaviorOptions when creating a Group using application permissions #978 [koenzomers - Koen Zomers]
+- Admin library: Ability to specify members on team site creation time #981 [LeHailender - Jens Haile]
 
 ### Changed
 
@@ -27,6 +31,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - Add MODERNCALENDAR  to the ViewType2 enum #939 [martinlingstuyl - Martin Lingstuyl]
 - Fix search sort direction with new POST approach #937 [lroos - Llewellyn Roos]
 - Ensure the Name property is loaded #955 [jansenbe - Bert Jansen]
+- VerifyProperties fails to detect other model types loaded via `QueryProperties` as dirty #961 [jansenbe - Bert Jansen]
+- Getting pages with ".aspx" in file name fails #974 [jansenbe - Bert Jansen]
+- Fix for failure at getting list items with a column entitled "Comments" #979 [jansenbe - Bert Jansen]
+- Adding missing property IsVisible property to AdaptiveCardAction #986 [koenzomers - Koen Zomers]
+- GetAvailableThemesAsync throws exception when 'isInverted' or 'name' values missing from JSON response #990 [jansenbe - Bert Jansen]
+- Lists.GetByTitle should be case insensitive #991 [jansenbe - Bert Jansen]
+- Making RichTextEditorInstanceId setable to enable consuming libraries to set inline images #997 [heinrich-ulbricht - Heinrich Ulbricht]
+- Escape characters in inline image `Caption`, `AltText` and `Link` properties to prevent malformed JSON #995 [jansenbe - Bert Jansen]
 
 ## [1.7.0]